requirements.txt
streamlit>=1.30
pytest>=7.0
sentence-transformers>=2.7,<3.0
transformers<4.46
numpy>=1.24
anthropic>=0.40
python-dotenv>=1.0
